Remove stale project artifacts

+++ src/main/java/com/example/taskmanager/Attachment.java
@@ -19,7 +19,7 @@ public class Attachment {
     private String metadata;
 
     @Column(name = "fileSize")
-    private int fileSize = 0;
+    private long fileSize = 0;
 
     @Column(name = "taskID")
     private Long taskID;
@@ -30,8 +30,8 @@ public class Attachment {
     public void setFileORLink(String fileORLink) { this.fileORLink = fileORLink; }
     public String getMetadata() { return metadata; }
     public void setMetadata(String metadata) { this.metadata = metadata; }
-    public int getFileSize() { return fileSize; }
-    public void setFileSize(int fileSize) { this.fileSize = fileSize; }
+    public long getFileSize() { return fileSize; }
+    public void setFileSize(long fileSize) { this.fileSize = fileSize; }
     public Long getTaskID() { return taskID; }
     public void setTaskID(Long taskID) { this.taskID = taskID; }
 }
